+++ app/styles/root.css
@@ -299,6 +299,12 @@ body {
   background-color: rgba(0, 0, 0, 0.4);
 }
 
+.fieldset {
+  border: none;
+  padding: 0;
+  margin: 0;
+}
+
 .popup {
   position: fixed;
   z-index: 2;
